Advancing the migration to pytest
