+++ conftest.py
@@ -8,4 +8,4 @@ pytest_plugins = ['aiida.manage.tests.pytest_fixtures']  # pylint: disable=inval
 
 @pytest.fixture(scope='function')
 def raspa_code(aiida_local_code_factory):  # pylint: disable=unused-argument
-    return aiida_local_code_factory("simulate", "raspa")
+    return aiida_local_code_factory("raspa", "simulate")
