+++ examples/files/IRMOF-1_eqeq.cif
