+++ examples/files/TCC1RS.cif
