+++ examples/files/pseudo_atoms.def
