+++ examples/simple_calculations/example_base.py
@@ -1,6 +1,5 @@
-#!/usr/bin/env python2
 # -*- coding: utf-8 -*-
-"""Restart from simple RASPA calculation."""
+"""Run simple RASPA calculation."""
 
 from __future__ import print_function
 from __future__ import absolute_import
@@ -10,44 +9,36 @@ import click
 
 from aiida.common import NotExistent
 from aiida.engine import run_get_pk, run
-from aiida.orm import Code, Dict, load_node
+from aiida.orm import Code, Dict
 from aiida.plugins import DataFactory
 
 # data objects
 CifData = DataFactory('cif')  # pylint: disable=invalid-name
 
 
-@click.command('cli')
-@click.argument('codelabel')
-@click.option('--previous_calc', '-p', required=True, type=int, help='PK of test_raspa_base.py calculation')
-@click.option('--submit', is_flag=True, help='Actually submit calculation')
-def main(codelabel, previous_calc, submit):
-    """Prepare and submit restart from simple RASPA calculation."""
-    try:
-        code = Code.get_from_string(codelabel)
-    except NotExistent:
-        print("The code '{}' does not exist".format(codelabel))
-        sys.exit(1)
+def example_base(raspa_code, submit=True):
+    """Prepare and submit simple RASPA calculation."""
 
     # parameters
     parameters = Dict(
         dict={
             "GeneralSettings": {
                 "SimulationType": "MonteCarlo",
-                "NumberOfCycles": 2000,
-                "NumberOfInitializationCycles": 2000,
-                "PrintEvery": 1000,
+                "NumberOfCycles": 400,
+                "NumberOfInitializationCycles": 200,
+                "PrintEvery": 200,
                 "Forcefield": "GenericMOFs",
                 "EwaldPrecision": 1e-6,
                 "CutOff": 12.0,
-                "HeliumVoidFraction": 0.149,
+                "WriteBinaryRestartFileEvery": 200,
             },
             "System": {
                 "tcc1rs": {
                     "type": "Framework",
                     "UnitCells": "1 1 1",
-                    "ExternalTemperature": 350.0,
-                    "ExternalPressure": 6e5,
+                    "HeliumVoidFraction": 0.149,
+                    "ExternalTemperature": 300.0,
+                    "ExternalPressure": 5e5,
                 },
             },
             "Component": {
@@ -62,18 +53,14 @@ def main(codelabel, previous_calc, submit):
         })
 
     # framework
-    framework = CifData(file=os.path.join(os.path.dirname(os.path.realpath(__file__)), 'files', 'TCC1RS.cif'))
-
-    # restart file
-    retrieved_parent_folder = load_node(previous_calc).outputs.retrieved
+    framework = CifData(file=os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', 'files', 'TCC1RS.cif'))
 
     # Contructing builder
-    builder = code.get_builder()
+    builder = raspa_code.get_builder()
     builder.framework = {
         "tcc1rs": framework,
     }
     builder.parameters = parameters
-    builder.retrieved_parent_folder = retrieved_parent_folder
     builder.metadata.options = {
         "resources": {
             "num_machines": 1,
@@ -86,7 +73,7 @@ def main(codelabel, previous_calc, submit):
     builder.metadata.store_provenance = True
 
     if submit:
-        print("Testing RASPA with simple input, restart ...")
+        print("Testing RASPA with simple input ...")
         res, pk = run_get_pk(builder)
         print("calculation pk: ", pk)
         print("Total Energy average (tcc1rs):", res['output_parameters'].dict.tcc1rs['general']['total_energy_average'])
@@ -101,7 +88,20 @@ def main(codelabel, previous_calc, submit):
     print("-----")
 
 
+@click.command('cli')
+@click.argument('codelabel')
+@click.option('--submit', is_flag=True, help='Actually submit calculation')
+def cli(codelabel, submit):
+    """Click interface"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
+    example_base(code, submit)
+
+
 if __name__ == '__main__':
-    main()  # pylint: disable=no-value-for-parameter
+    cli()  # pylint: disable=no-value-for-parameter
 
 # EOF
