+++ examples/simple_calculations/example_base_restart.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python2
 # -*- coding: utf-8 -*-
-"""Run RASPA calculation to compute Henry coefficient."""
+"""Restart from simple RASPA calculation."""
+
 from __future__ import print_function
 from __future__ import absolute_import
 import os
@@ -9,62 +10,61 @@ import click
 
 from aiida.common import NotExistent
 from aiida.engine import run_get_pk, run
+from aiida.orm import Code, Dict, load_node
 from aiida.plugins import DataFactory
-from aiida.orm import Code, Dict
 
 # data objects
 CifData = DataFactory('cif')  # pylint: disable=invalid-name
 
 
-@click.command('cli')
-@click.argument('codelabel')
-@click.option('--submit', is_flag=True, help='Actually submit calculation')
-def main(codelabel, submit):
-    """Prepare and submit simple RASPA calculation to compute Henry coefficient."""
-    try:
-        code = Code.get_from_string(codelabel)
-    except NotExistent:
-        print("The code '{}' does not exist".format(codelabel))
-        sys.exit(1)
+def example_base_restart(raspa_code, previous_calc, submit=True):
+    """Prepare and submit restart from simple RASPA calculation."""
 
     # parameters
     parameters = Dict(
         dict={
             "GeneralSettings": {
                 "SimulationType": "MonteCarlo",
-                "NumberOfCycles": 400,
-                "PrintEvery": 200,
+                "NumberOfCycles": 2000,
+                "NumberOfInitializationCycles": 2000,
+                "PrintEvery": 1000,
                 "Forcefield": "GenericMOFs",
                 "EwaldPrecision": 1e-6,
                 "CutOff": 12.0,
+                "HeliumVoidFraction": 0.149,
             },
             "System": {
                 "tcc1rs": {
                     "type": "Framework",
                     "UnitCells": "1 1 1",
-                    "HeliumVoidFraction": 0.149,
-                    "ExternalTemperature": 300.0,
-                }
+                    "ExternalTemperature": 350.0,
+                    "ExternalPressure": 6e5,
+                },
             },
             "Component": {
                 "methane": {
                     "MoleculeDefinition": "TraPPE",
-                    "WidomProbability": 1.0,
+                    "TranslationProbability": 0.5,
+                    "ReinsertionProbability": 0.5,
+                    "SwapProbability": 1.0,
                     "CreateNumberOfMolecules": 0,
                 }
             },
         })
 
     # framework
-    pwd = os.path.dirname(os.path.realpath(__file__))
-    framework = CifData(file=pwd + '/files/TCC1RS.cif')
+    framework = CifData(file=os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', 'files', 'TCC1RS.cif'))
+
+    # restart file
+    retrieved_parent_folder = load_node(previous_calc).outputs.retrieved
 
     # Contructing builder
-    builder = code.get_builder()
+    builder = raspa_code.get_builder()
     builder.framework = {
         "tcc1rs": framework,
     }
     builder.parameters = parameters
+    builder.retrieved_parent_folder = retrieved_parent_folder
     builder.metadata.options = {
         "resources": {
             "num_machines": 1,
@@ -77,23 +77,36 @@ def main(codelabel, submit):
     builder.metadata.store_provenance = True
 
     if submit:
-        print("Testing RASPA on computing Henry coefficient ...")
+        print("Testing RASPA with simple input, restart ...")
         res, pk = run_get_pk(builder)
         print("calculation pk: ", pk)
-        print("Average Henry coefficient (methane in tcc1rs):",
-              res['output_parameters'].dict.tcc1rs['components']['methane']['henry_coefficient_average'])
+        print("Total Energy average (tcc1rs):", res['output_parameters'].dict.tcc1rs['general']['total_energy_average'])
         print("OK, calculation has completed successfully")
     else:
         print("Generating test input ...")
         builder.metadata.dry_run = True
         builder.metadata.store_provenance = False
         run(builder)
-        print("submission test successful")
+        print("Submission test successful")
         print("In order to actually submit, add '--submit'")
     print("-----")
 
 
+@click.command('cli')
+@click.argument('codelabel')
+@click.option('--previous_calc', '-p', required=True, type=int, help='PK of example_base.py calculation')
+@click.option('--submit', is_flag=True, help='Actually submit calculation')
+def cli(codelabel, previous_calc, submit):
+    """Click interface"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
+    example_base_restart(code, previous_calc, submit)
+
+
 if __name__ == '__main__':
-    main()  # pylint: disable=no-value-for-parameter
+    cli()  # pylint: disable=no-value-for-parameter
 
 # EOF
