+++ examples/simple_calculations/example_framework_box.py
@@ -1,5 +1,7 @@
+#!/usr/bin/env python2
 # -*- coding: utf-8 -*-
-"""Run RASPA calculation with blocked pockets."""
+"""Run simple RASPA calculation."""
+
 from __future__ import print_function
 from __future__ import absolute_import
 import os
@@ -8,23 +10,16 @@ import click
 
 from aiida.common import NotExistent
 from aiida.engine import run_get_pk, run
-from aiida.orm import Code, Dict, SinglefileData
+from aiida.orm import Code, Dict
 from aiida.plugins import DataFactory
 
 # data objects
 CifData = DataFactory('cif')  # pylint: disable=invalid-name
 
 
-@click.command('cli')
-@click.argument('codelabel')
-@click.option('--submit', is_flag=True, help='Actually submit calculation')
-def main(codelabel, submit):
-    """Prepare and submit RASPA calculation with blocked pockets."""
-    try:
-        code = Code.get_from_string(codelabel)
-    except NotExistent:
-        print("The code '{}' does not exist".format(codelabel))
-        sys.exit(1)
+def example_framework_box(raspa_code, submit=True):
+    """"Test RASPA with framework and box."""
+
     # parameters
     parameters = Dict(
         dict={
@@ -32,19 +27,25 @@ def main(codelabel, submit):
                 "SimulationType": "MonteCarlo",
                 "NumberOfCycles": 400,
                 "NumberOfInitializationCycles": 200,
-                "PrintEvery": 200,
+                "PrintEvery": 100,
                 "Forcefield": "GenericMOFs",
                 "EwaldPrecision": 1e-6,
-                "CutOff": 12.0,
+                "WriteBinaryRestartFileEvery": 200,
             },
             "System": {
                 "tcc1rs": {
                     "type": "Framework",
                     "UnitCells": "1 1 1",
-                    "HeliumVoidFraction": 0.149,
                     "ExternalTemperature": 300.0,
                     "ExternalPressure": 5e5,
-                }
+                    "HeliumVoidFraction": 0.149,
+                },
+                "box_25_angstroms": {
+                    "type": "Box",
+                    "BoxLengths": "25 25 25",
+                    "ExternalTemperature": 300.0,
+                    "ExternalPressure": 5e5,
+                },
             },
             "Component": {
                 "methane": {
@@ -52,26 +53,22 @@ def main(codelabel, submit):
                     "TranslationProbability": 0.5,
                     "ReinsertionProbability": 0.5,
                     "SwapProbability": 1.0,
-                    "CreateNumberOfMolecules": 0,
-                    "BlockPocketsFileName": "block_tcc1rs_methane",
+                    "CreateNumberOfMolecules": {
+                        "tcc1rs": 1,
+                        "box_25_angstroms": 2,
+                    }
                 }
             },
         })
 
-    # framework
     pwd = os.path.dirname(os.path.realpath(__file__))
-    framework = CifData(file=os.path.join(pwd, 'files', 'TCC1RS.cif'))
-
-    # block pocket
-    block_pocket_node = SinglefileData(file=os.path.join(pwd, 'files', 'block_pocket.block')).store()
+    # framework
+    framework = CifData(file=os.path.join(pwd, '..', 'files', 'TCC1RS.cif'))
 
     # Contructing builder
-    builder = code.get_builder()
+    builder = raspa_code.get_builder()
     builder.framework = {
         "tcc1rs": framework,
-    }
-    builder.block_pocket = {
-        "block_tcc1rs_methane": block_pocket_node,
     }
     builder.parameters = parameters
     builder.metadata.options = {
@@ -86,10 +83,12 @@ def main(codelabel, submit):
     builder.metadata.store_provenance = True
 
     if submit:
-        print("Testing RASPA with block pockets ...")
+        print("Testing RASPA with framework and box ...")
         res, pk = run_get_pk(builder)
         print("calculation pk: ", pk)
         print("Total Energy average (tcc1rs):", res['output_parameters'].dict.tcc1rs['general']['total_energy_average'])
+        print("Total Energy average (box_25_angstroms):",
+              res['output_parameters'].dict.box_25_angstroms['general']['total_energy_average'])
         print("OK, calculation has completed successfully")
     else:
         print("Generating test input ...")
@@ -98,9 +97,23 @@ def main(codelabel, submit):
         run(builder)
         print("Submission test successful")
         print("In order to actually submit, add '--submit'")
+    print("-----")
+
+
+@click.command('cli')
+@click.argument('codelabel')
+@click.option('--submit', is_flag=True, help='Actually submit calculation')
+def cli(codelabel, submit):
+    """Click interface"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
+    example_framework_box(code, submit)
 
 
 if __name__ == '__main__':
-    main()  # pylint: disable=no-value-for-parameter
+    cli()  # pylint: disable=no-value-for-parameter
 
 # EOF
