+++ examples/simple_calculations/example_gemc_single_comp.py
@@ -1,32 +1,18 @@
-#!/usr/bin/env python2
 # -*- coding: utf-8 -*-
-"""Run simple RASPA calculation."""
+"""Run RASPA single-component GEMC calculation"""
 
 from __future__ import print_function
 from __future__ import absolute_import
-import os
 import sys
 import click
 
 from aiida.common import NotExistent
 from aiida.engine import run_get_pk, run
 from aiida.orm import Code, Dict
-from aiida.plugins import DataFactory
-
-# data objects
-CifData = DataFactory('cif')  # pylint: disable=invalid-name
 
 
-@click.command('cli')
-@click.argument('codelabel')
-@click.option('--submit', is_flag=True, help='Actually submit calculation')
-def main(codelabel, submit):
-    """Prepare and submit simple RASPA calculation."""
-    try:
-        code = Code.get_from_string(codelabel)
-    except NotExistent:
-        print("The code '{}' does not exist".format(codelabel))
-        sys.exit(1)
+def example_gemc_single_comp(raspa_code, submit=True):
+    """Prepare and submit RASPA calculation with components mixture."""
 
     # parameters
     parameters = Dict(
@@ -35,48 +21,42 @@ def main(codelabel, submit):
                 "SimulationType": "MonteCarlo",
                 "NumberOfCycles": 400,
                 "NumberOfInitializationCycles": 200,
-                "PrintEvery": 100,
+                "PrintEvery": 200,
                 "Forcefield": "GenericMOFs",
                 "EwaldPrecision": 1e-6,
-                "WriteBinaryRestartFileEvery": 200,
+                "CutOff": 12.0,
+                "GibbsVolumeChangeProbability": 0.1,
             },
             "System": {
-                "tcc1rs": {
-                    "type": "Framework",
-                    "UnitCells": "1 1 1",
-                    "ExternalTemperature": 300.0,
-                    "ExternalPressure": 5e5,
-                    "HeliumVoidFraction": 0.149,
-                },
-                "box_25_angstroms": {
+                "box_one": {
                     "type": "Box",
                     "BoxLengths": "25 25 25",
+                    "BoxAngles": "90 90 90",
                     "ExternalTemperature": 300.0,
-                    "ExternalPressure": 5e5,
                 },
+                "box_two": {
+                    "type": "Box",
+                    "BoxLengths": "25 25 25",
+                    "BoxAngles": "90 90 90",
+                    "ExternalTemperature": 300.0,
+                }
             },
             "Component": {
                 "methane": {
                     "MoleculeDefinition": "TraPPE",
-                    "TranslationProbability": 0.5,
-                    "ReinsertionProbability": 0.5,
-                    "SwapProbability": 1.0,
+                    "TranslationProbability": 1.0,
+                    "ReinsertionProbability": 1.0,
+                    "GibbsSwapProbability": 1.0,
                     "CreateNumberOfMolecules": {
-                        "tcc1rs": 1,
-                        "box_25_angstroms": 2,
-                    }
-                }
+                        "box_one": 150,
+                        "box_two": 150,
+                    },
+                },
             },
         })
 
-    # framework
-    framework = CifData(file=os.path.join(os.path.dirname(os.path.realpath(__file__)), 'files', 'TCC1RS.cif'))
-
     # Contructing builder
-    builder = code.get_builder()
-    builder.framework = {
-        "tcc1rs": framework,
-    }
+    builder = raspa_code.get_builder()
     builder.parameters = parameters
     builder.metadata.options = {
         "resources": {
@@ -90,12 +70,13 @@ def main(codelabel, submit):
     builder.metadata.store_provenance = True
 
     if submit:
-        print("Testing RASPA with framework and box ...")
+        print("Testing RASPA GEMC with methane ...")
         res, pk = run_get_pk(builder)
         print("calculation pk: ", pk)
-        print("Total Energy average (tcc1rs):", res['output_parameters'].dict.tcc1rs['general']['total_energy_average'])
-        print("Total Energy average (box_25_angstroms):",
-              res['output_parameters'].dict.box_25_angstroms['general']['total_energy_average'])
+        print("Total Energy average (box_one):",
+              res['output_parameters'].dict.box_one['general']['total_energy_average'])
+        print("Total Energy average (box_two):",
+              res['output_parameters'].dict.box_two['general']['total_energy_average'])
         print("OK, calculation has completed successfully")
     else:
         print("Generating test input ...")
@@ -107,7 +88,20 @@ def main(codelabel, submit):
     print("-----")
 
 
+@click.command('cli')
+@click.argument('codelabel')
+@click.option('--submit', is_flag=True, help='Actually submit calculation')
+def cli(codelabel, submit):
+    """Click interface"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
+    example_gemc_single_comp(code, submit)
+
+
 if __name__ == '__main__':
-    main()  # pylint: disable=no-value-for-parameter
+    cli()  # pylint: disable=no-value-for-parameter
 
 # EOF
