+++ examples/workchains/example_base_workchain_gcmc_2comp.py
@@ -13,18 +13,10 @@ from aiida.orm import CifData, Code, Dict, SinglefileData
 from aiida_raspa.workchains import RaspaBaseWorkChain
 
 
-@click.command('cli')
-@click.argument('codelabel')
-def main(codelabel):
-    """Run base workchain"""
+def example_base_workchain_gcmc(raspa_code):
+    """Run base workchain for GCMC calculations with 2 components."""
 
     # pylint: disable=no-member
-
-    try:
-        code = Code.get_from_string(codelabel)
-    except NotExistent:
-        print("The code '{}' does not exist".format(codelabel))
-        sys.exit(1)
 
     print("Testing RASPA Xenon:Krypton GCMC through RaspaBaseWorkChain ...")
 
@@ -73,19 +65,17 @@ def main(codelabel):
 
     # framework
     pwd = os.path.dirname(os.path.realpath(__file__))
-    structure = CifData(file=os.path.join(pwd, '..', 'simple_calculations', 'files', 'IRMOF-1.cif'))
+    structure = CifData(file=os.path.join(pwd, '..', 'files', 'IRMOF-1.cif'))
     structure_label = "irmof_1"
 
-    block_pocket_node1 = SinglefileData(
-        file=os.path.join(pwd, '..', 'simple_calculations', 'files', 'IRMOF-1_test.block')).store()
-    block_pocket_node2 = SinglefileData(
-        file=os.path.join(pwd, '..', 'simple_calculations', 'files', 'IRMOF-1_test.block')).store()
+    block_pocket_node1 = SinglefileData(file=os.path.join(pwd, '..', 'files', 'IRMOF-1_test.block')).store()
+    block_pocket_node2 = SinglefileData(file=os.path.join(pwd, '..', 'files', 'IRMOF-1_test.block')).store()
 
     # Constructing builder
     builder = RaspaBaseWorkChain.get_builder()
 
     # Specifying the code
-    builder.raspa.code = code
+    builder.raspa.code = raspa_code
 
     # Specifying the framework
     builder.raspa.framework = {
@@ -119,7 +109,19 @@ def main(codelabel):
     run(builder)
 
 
+@click.command('cli')
+@click.argument('codelabel')
+def cli(codelabel):
+    """Click interface"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
+    example_base_workchain_gcmc(code)
+
+
 if __name__ == '__main__':
-    main()  # pylint: disable=no-value-for-parameter
+    cli()  # pylint: disable=no-value-for-parameter
 
 # EOF
